Handle deleting symlinks differently on Windows

+++ java/meterpreter/stdapi/src/main/java/com/metasploit/meterpreter/stdapi/stdapi_fs_delete_dir_V1_7.java
@@ -7,7 +7,11 @@ import java.nio.file.Files;
 public class stdapi_fs_delete_dir_V1_7 extends stdapi_fs_delete_dir {
     @Override
     protected boolean deleteSymlink(File file) throws IOException {
-        Files.delete(file.toPath());
-        return true;
+        String osName = System.getProperty("os.name");
+        if (osName != null && osName.toLowerCase().contains("windows")) {
+            Files.delete(file.toPath());
+            return true;
+        }
+        return file.delete();
     }
 }
